+++ CHANGELOG.md
@@ -7,8 +7,14 @@ Features:
 * Expand set of allowed protocols to include `tel:` and `line:`. [#104, #147]
 * Expand set of allowed CSS functions. [related to #122]
 * Allow greater precision in shorthand CSS values. [#149] (Thanks, @danfstucky!)
-
-
+* Update terminology
+  - Deprecate `Loofah::Helpers::ActionView.white_list_sanitizer`,
+    please use `Loofah::Helpers::ActionView.safe_list_sanitizer` instead.
+  - Deprecate `Loofah::Helpers::ActionView::WhiteListSanitizer`,
+    please use `Loofah::Helpers::ActionView::SafeListSanitizer` instead.
+  - Deprecate `Loofah::HTML5::WhiteList`,
+    please use `Loofah::HTML5::SafeList` instead.
+    
 ## 2.2.3 / 2018-10-30
 
 ### Security